pyproject.toml
[project]
name = "x-post"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["requests", "python-dotenv"]

[project.scripts]
x-post = "x.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.hatch.build.targets.wheel]
packages = ["src/x"]
